requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.115.13
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.5.1
numpy==2.3.1
openai==1.91.0
orjson==3.10.18
packaging==24.2
pillow==11.2.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
together==1.5.17
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
yarl==1.20.1
zstandard==0.23.0
